data-center/actor/requirements.txt
fastapi==0.105.0
paho-mqtt==1.6.1
pydantic==2.5.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
python-dotenv